requirements.txt
Django==4.2
psycopg2-binary==2.9.6  # PostgreSQL adapter
google-api-python-client==2.85.0  # Google Calendar API
google-auth # Removed version pin to allow compatible version installation
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
requests==2.28.1  # Para la integración con ManyChat
httpx>=0.27.0  # Cliente HTTP asíncrono para MCP API
djangorestframework==3.14.0  # Django REST Framework
gunicorn
openai
dj-database-url>=0.5.0
whitenoise
django-cors-headers
xlwt==1.3.0  # Para exportación a Excel
openpyxl==3.1.2  # Para importar desde Excel
Pillow>=9.0.0     # For ImageField support
phonenumbers>=8.12.0 # For phone number validation/formatting
google-cloud-storage>=2.0.0 # For Google Cloud Storage
django-storages>=1.14.0 # For Django storage backends (Removed [google] extra)
django-solo>=2.1.0 # For singleton configuration models
WeasyPrint>=57.1   # Para generar PDF
